Rework GroundHandler as a 2d Character controller that replaces the previous PenguinController
