+++ Assets/PenguinQuest/Code/Controllers/AlwaysOnComponents/CharacterController2D.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ba06e4e983cbe944687dd3516a15755e
+guid: 01e9c817aeb9b3c449ab72edec42564a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
