--- Assets/PenguinQuest/Code/Controllers/PenguinController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47ecfa2bad6a73c409012eced7381d03
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
